ranges can have floating point increments
git-svn-id: https://svn.ci.uchicago.edu/svn/vdl2/trunk@1043 e2bb083e-7f23-0410-b3a8-8253ac9ef6d8

+++ src/org/griphyn/vdl/karajan/lib/Range.java
@@ -14,7 +14,7 @@ import org.griphyn.vdl.mapping.RootArrayDataNode;
 public class Range extends VDLFunction {
 	public static final SwiftArg PA_FROM = new SwiftArg.Positional("from");
 	public static final SwiftArg PA_TO = new SwiftArg.Positional("to");
-	public static final SwiftArg OA_STEP = new SwiftArg.Optional("step", new Double(1), "int");
+	public static final SwiftArg OA_STEP = new SwiftArg.Optional("step", new Double(1),"float");
 
 	static {
 		setArguments(Range.class, new Arg[] { PA_FROM, PA_TO, OA_STEP });
